image-processing/requirements.txt
scikit-image
numpy
matplotlib
Pillow
pypng
opencv-python
pytest-pep8
